updated main view styles, to change later and move styles to login page
