+++ apps/web/src/components/pages/MainTemplate.tsx
@@ -5,7 +5,7 @@ import Navbar from "../ui/Navbar";
 const MainTemplate = () => {
 
     return (
-        <div className="w-screen h-screen m-0 p-0 overflow-hidden">
+        <div className="w-screen h-screen m-0 p-0 overflow-hidden bg-pastel-bg">
             <Navbar />
             <EmployeeHubProvider>
                 <Outlet />
