+++ apps/web/src/components/ui/Grid.tsx
@@ -3,7 +3,7 @@ import Card from "./Card"
 
 const Grid = ({ users }: GridProps) => {
     return (
-        <div className="text-gray-200">
+        <div className="text-gray-900">
             <h1 className="font-medium text-lg p-8">Pracownicy</h1>
             <div className="h-full flex flex-col p-5 items-center">
                 <div className="w-2/3">
