+++ apps/web/src/providers/ContextProvider.tsx
@@ -1,6 +1,6 @@
 import React, { PropsWithChildren, createContext } from 'react';
 import { IUser } from '../types/types.user';
-import { IEmployeeHubContextType } from '../types/types.employeeHubContext';
+import { IEmployeeHubContextType } from '../types/providers/types.employeeHubContext';
 
 
 export const EmployeeHubContext = createContext<IEmployeeHubContextType>({
