+++ apps/web/src/types/providers/types.employeeHubContext.tsx
@@ -1,4 +1,4 @@
-import { IUser } from "./types.user";
+import { IUser } from "../types.user";
 
 export interface IEmployeeHubContextType {
     users:IUser[],
